Cleaning mess
- refactoring of includes
- make Policy and RuleSet public
- move all inherited RuleSets into Daemon/
