+++ src/Library/public/usbguard/RuleSet.hpp
@@ -1,5 +1,5 @@
 //
-// Copyright (C) 2017 Red Hat, Inc.
+// Copyright (C) 2015 Red Hat, Inc.
 //
 // This program is free software; you can redistribute it and/or modify
 // it under the terms of the GNU General Public License as published by
@@ -14,41 +14,37 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 //
-// Authors: Daniel Kopecek <dkopecek@redhat.com>
-//          Radovan Sroka <rsroka@redhat.com>
+// Authors: Radovan Sroka <rsroka@redhat.com>
 //
 #pragma once
-#ifdef HAVE_BUILD_CONFIG_H
-  #include <build-config.h>
-#endif
 
 #include "usbguard/Typedefs.hpp"
+#include "usbguard/Rule.hpp"
 
-#include <string>
-#include <vector>
-
-#include "../RuleSet.hpp"
+#include <istream>
+#include <ostream>
+#include <mutex>
 
 namespace usbguard
 {
-  class DLL_PUBLIC Policy
+  class Interface;
+  class DLL_PUBLIC RuleSet
   {
   public:
-    enum class EventType {
-      Insert = 1,
-      Update = 2,
-      Remove = 3
-    };
 
-    Policy();
+    RuleSet(Interface* const interface_ptr);
+    RuleSet(const RuleSet& rhs);
+    const RuleSet& operator=(const RuleSet& rhs);
 
-    void setRuleSet(std::shared_ptr<RuleSet> ptr);
-    std::shared_ptr<RuleSet> getRuleSet();
+    virtual void load();
+    virtual void save();
+
+    void serialize(std::ostream& stream) const;
 
     void setDefaultTarget(Rule::Target target);
     Rule::Target getDefaultTarget() const;
     void setDefaultAction(const std::string& action);
-    uint32_t appendRule(const Rule& rule, uint32_t parent_id = Rule::LastID);
+    uint32_t appendRule(const Rule& rule, uint32_t parent_id = Rule::LastID, bool lock = true);
     uint32_t upsertRule(const Rule& match_rule, const Rule& new_rule, bool parent_insensitive = false);
     std::shared_ptr<Rule> getRule(uint32_t id);
     bool removeRule(uint32_t id);
@@ -57,10 +53,33 @@ namespace usbguard
     uint32_t assignID(std::shared_ptr<Rule> rule);
     uint32_t assignID();
 
-    static std::string eventTypeToString(EventType event);
-  private:
+    void setWritable();
+    void clearWritable();
+    bool isWritable();
 
-    std::shared_ptr<RuleSet> _ruleset_ptr;
+  protected:
+    mutable std::mutex _io_mutex; /* mutex for load/save */
+    mutable std::mutex _op_mutex; /* mutex for operations on the rule set */
+
+    bool _writable;
+
+    Interface* const _interface_ptr;
+    Rule::Target _default_target;
+    std::string _default_action;
+    Atomic<uint32_t> _id_next;
+    std::vector<std::shared_ptr<Rule>> _rules;
+  };
+
+  class DLL_PUBLIC RuleSetAbstract : public RuleSet
+  {
+  public:
+    RuleSetAbstract(Interface* const interface_ptr)
+      : RuleSet(interface_ptr) {};
+    RuleSetAbstract(const RuleSetAbstract& rhs)
+      : RuleSet(rhs._interface_ptr) {};
+
+    virtual void load() = 0;
+    virtual void save() = 0;
   };
 } /* namespace usbguard */
